extract and modify setDimensions
